gen: Move the imp submodule generation to an imp.rs module
